+++ NEWS.md
@@ -1,4 +1,6 @@
-# MAIVE (development version)
+# MAIVE 0.5.0
+
+*Released: 2026-09-14*
 
 ## Bug Fixes